chore: refactor ship data into one file
