+++ packages/client/src/layers/react/components/OverviewComponents/EnemyShip.tsx
@@ -1,14 +1,9 @@
 import { GodID } from "@latticexyz/network";
 import { EntityIndex, getComponentValue, getComponentValueStrict } from "@latticexyz/recs";
 import { map, merge, of } from "rxjs";
-import { SailPositions } from "../../../../types";
-import { getShipSprite, ShipImages } from "../../../../utils/ships";
-import { ShipAttributeTypes } from "../../../phaser/constants";
 import { registerUIComponent } from "../../engine";
-import { BoxImage, Container, InternalContainer } from "../../styles/global";
-import HullHealth from "./HullHealth";
-import ShipAttribute from "./ShipAttribute";
-import ShipDamage from "./ShipDamage";
+import { Container, InternalContainer } from "../../styles/global";
+import { ShipCard } from "../YourShips/ShipCard";
 
 export function registerEnemyShip() {
   registerUIComponent(
@@ -71,122 +66,29 @@ export function registerEnemyShip() {
       ).pipe(
         map(() => {
           return {
-            world,
-            SelectedShip,
-            Position,
-            Rotation,
-            OwnedBy,
-            Health,
-            SailPosition,
-            CrewCount,
-            DamagedMast,
-            Firepower,
-            Leak,
-            Ship,
-            OnFire,
-            Name,
-            Player,
-            connectedAddress,
-            camera,
-            positions,
+            layers,
             getPlayerEntity,
+            connectedAddress,
+            OwnedBy,
+            SelectedShip,
+            world,
           };
         })
       );
     },
-    (props) => {
-      const {
-        SelectedShip,
-        Rotation,
-        Name,
-        Position,
-        Health,
-        Player,
-        CrewCount,
-        Firepower,
-        SailPosition,
-        DamagedMast,
-        OwnedBy,
-        OnFire,
-        Leak,
-        world,
-        connectedAddress,
-        getPlayerEntity,
-      } = props;
-
-      const playerEntity = getPlayerEntity(connectedAddress.get());
-      if (!playerEntity || !getComponentValue(Player, playerEntity)) return null;
-
+    ({ layers, getPlayerEntity, connectedAddress, OwnedBy, SelectedShip, world }) => {
       const GodEntityIndex: EntityIndex = world.entityToIndex.get(GodID) || (0 as EntityIndex);
 
-      const selectedShip = getComponentValue(SelectedShip, GodEntityIndex)?.value as EntityIndex | undefined;
-      if (!selectedShip) return null;
+      const ship = getComponentValue(SelectedShip, GodEntityIndex)?.value as EntityIndex | undefined;
+      if (!ship) return null;
 
-      const ownerEntity = getPlayerEntity(getComponentValueStrict(OwnedBy, selectedShip).value);
+      const playerEntity = getPlayerEntity(connectedAddress.get());
+      const ownerEntity = getPlayerEntity(getComponentValueStrict(OwnedBy, ship).value);
       if (!ownerEntity || playerEntity == ownerEntity) return null;
-
-      const sailPosition = getComponentValueStrict(SailPosition, selectedShip).value;
-      const rotation = getComponentValueStrict(Rotation, selectedShip).value;
-      const position = getComponentValueStrict(Position, selectedShip);
-      const health = getComponentValueStrict(Health, selectedShip).value;
-      const crewCount = getComponentValueStrict(CrewCount, selectedShip).value;
-      const firepower = getComponentValueStrict(Firepower, selectedShip).value;
-      const onFire = getComponentValue(OnFire, selectedShip)?.value;
-      const leak = getComponentValue(Leak, selectedShip)?.value;
-      const damagedMast = getComponentValue(DamagedMast, selectedShip)?.value;
-      const ownerName = getComponentValue(Name, ownerEntity)?.value;
-
       return (
         <Container style={{ justifyContent: "flex-end" }}>
           <InternalContainer style={{ gap: "24px" }}>
-            <div style={{ display: "flex", borderRadius: "6px", width: "100%", height: "100%" }}>
-              <div
-                style={{
-                  flex: 2,
-                  display: "flex",
-                  flexDirection: "column",
-                  height: "100%",
-                  position: "relative",
-                  maxWidth: "140px",
-                  minWidth: "80px",
-                }}
-              >
-                <span style={{ fontSize: "1.5rem", lineHeight: "2.5rem" }}>HMS {selectedShip}</span>
-                {ownerName && <span style={{ lineHeight: "1.5rem", fontSize: "1rem" }}>{ownerName}</span>}
-                <span style={{ lineHeight: "1.5rem", fontSize: "1rem" }}>
-                  ({position.x}, {position.y})
-                </span>
-                <BoxImage>
-                  <img
-                    src={ShipImages[getShipSprite(playerEntity, ownerEntity, health)]}
-                    style={{
-                      objectFit: "scale-down",
-                      left: "50%",
-                      position: "absolute",
-                      top: "50%",
-                      margin: "auto",
-                      transform: `rotate(${rotation - 90}deg) translate(-50%,-50%)`,
-                      transformOrigin: `top left`,
-                      maxWidth: "50px",
-                    }}
-                  />
-                </BoxImage>
-              </div>
-              <div style={{ flex: 3, display: "flex", flexDirection: "column" }}>
-                <HullHealth health={health} />
-                <div style={{ display: "flex", width: "100%" }}>
-                  <ShipAttribute attributeType={ShipAttributeTypes.Crew} attribute={crewCount} />
-                  <ShipAttribute attributeType={ShipAttributeTypes.Firepower} attribute={firepower} />
-                  <ShipAttribute attributeType={ShipAttributeTypes.Sails} attribute={SailPositions[sailPosition]} />
-                </div>
-                <div style={{ display: "flex" }}>
-                  {damagedMast && <ShipDamage message="mast broken" amountLeft={damagedMast} />}
-                  {onFire && <ShipDamage message="on fire" amountLeft={onFire} />}
-                  {leak && <ShipDamage message="leaking" />}
-                  {sailPosition == 0 && <ShipDamage message="sails torn" />}
-                </div>
-              </div>
-            </div>
+            <ShipCard layers={layers} ship={ship} />
           </InternalContainer>
         </Container>
       );
